+++ packages/graphql-codegen-cli/tests/test-files/loaders/module-exports.js
@@ -1,0 +1,17 @@
+var graphql = require('graphql');
+
+var schema = new graphql.GraphQLSchema({
+  query: new graphql.GraphQLObjectType({
+    name: 'RootQueryType',
+    fields: {
+      hello: {
+        type: graphql.GraphQLString,
+        resolve() {
+          return 'world';
+        }
+      }
+    }
+  })
+});
+
+module.exports = schema;